Removed deprecated usage of internal class ArrayChannel
